+++ packages/@tinacms/cli/src/cmds/baseCmds.ts
@@ -21,16 +21,20 @@ import {
   tinaSetup,
 } from './init'
 
+import 'dotenv/config'
 import { Command } from '../command'
 import { chain } from '../middleware'
 import chalk from 'chalk'
-import { compile } from './compile'
+import { compileSchema, compileClient } from './compile'
 import { logger } from '../logger'
 import { startServer } from './start-server'
+import { waitForDB } from './waitForDB'
+import { startSubprocess } from './startSubprocess'
 
 export const CMD_GEN_TYPES = 'schema:types'
 export const CMD_START_SERVER = 'server:start'
 export const CMD_COMPILE_MODELS = 'schema:compile'
+export const CMD_WAIT_FOR_DB = 'server:waitForDB'
 export const INIT = 'init'
 export const AUDIT = 'audit'
 
@@ -87,6 +91,10 @@ const tinaCloudMediaStore = {
   description:
     'Automatically pushes updates from GitHub to the Tina Cloud Media Store',
 }
+const developmentOption = {
+  name: '--dev',
+  description: 'Uses NODE_ENV=development when compiling client and schema',
+}
 
 export const baseCmds: Command[] = [
   {
@@ -102,14 +110,28 @@ export const baseCmds: Command[] = [
       noTelemetryOption,
       watchFileOption,
       verboseOption,
+      developmentOption,
     ],
-    action: (options) => chain([startServer], options),
+    action: (options) => chain([startServer, startSubprocess], options),
+  },
+  {
+    command: CMD_WAIT_FOR_DB,
+    description: 'Wait for DB to finish indexing, start subprocess',
+    options: [
+      subCommand,
+      experimentalDatalayer,
+      noTelemetryOption,
+      verboseOption,
+      developmentOption,
+    ],
+    action: (options) =>
+      chain([compileClient, waitForDB, startSubprocess], options),
   },
   {
     command: CMD_COMPILE_MODELS,
     description: 'Compile schema into static files for the server',
     options: [experimentalDatalayer, tinaCloudMediaStore, noTelemetryOption],
-    action: (options) => chain([compile], options),
+    action: (options) => chain([compileSchema], options),
   },
   {
     command: CMD_GEN_TYPES,
@@ -139,7 +161,7 @@ export const baseCmds: Command[] = [
           initTina,
           installDeps,
           async (_ctx, next, options) => {
-            await compile(_ctx, next, options)
+            await compileSchema(_ctx, next, options)
             next()
           },
           attachSchema,
@@ -163,7 +185,7 @@ export const baseCmds: Command[] = [
             next()
           },
           async (_ctx, next) => {
-            await compile(_ctx, next, options)
+            await compileSchema(_ctx, next, options)
             next()
           },
           attachSchema,