+++ packages/@tinacms/graphql/src/resolver/index.ts
@@ -34,6 +34,10 @@ import type { GraphQLConfig } from '../types'
 import { TinaError } from './error'
 import { FilterCondition, makeFilterChain } from '@tinacms/datalayer'
 import { collectConditionsForField, resolveReferences } from './filter-utils'
+import {
+  resolveMediaRelativeToCloud,
+  resolveMediaCloudToRelative,
+} from './media-utils'
 
 interface ResolverConfig {
   config?: GraphQLConfig
@@ -484,11 +488,6 @@ export class Resolver {
     }
   }
 
-  public getDocumentsForCollection = async (collectionName: string) => {
-    const collection = this.tinaSchema.getCollection(collectionName)
-    return this.database.store.glob(collection.path, this.getDocument)
-  }
-
   private referenceResolver = async (
     filter: Record<string, object>,
     fieldDefinition: ReferenceTypeWithNamespace
@@ -577,7 +576,14 @@ export class Resolver {
     let edges
     let pageInfo
 
-    if (args.filter || args.sort) {
+    // See if we are using the data layer
+    const useDataLayer = Boolean(
+      this.tinaSchema?.config?.meta?.flags?.find(
+        (x) => x === 'experimentalData'
+      )
+    )
+
+    if (useDataLayer) {
       let conditions: FilterCondition[]
       if (args.filter) {
         if (collection.fields) {
@@ -668,15 +674,19 @@ export class Resolver {
         case 'string':
         case 'boolean':
         case 'number':
-        case 'image':
           accum[fieldName] = fieldValue
+          break
+        case 'image':
+          accum[fieldName] = resolveMediaCloudToRelative(
+            fieldValue as string,
+            this.config
+          )
           break
         case 'object':
           accum[fieldName] = this.buildObjectMutations(fieldValue, field)
           break
         case 'rich-text':
-          field
-          accum[fieldName] = stringifyMDX(fieldValue, field)
+          accum[fieldName] = stringifyMDX(fieldValue, field, this.config)
           break
         case 'reference':
           accum[fieldName] = fieldValue
@@ -715,21 +725,14 @@ export class Resolver {
         accumulator[field.name] = value
         break
       case 'image':
-        if (this.config) {
-          if (this.config.useRelativeMedia === true) {
-            accumulator[field.name] = value
-          } else {
-            accumulator[
-              field.name
-            ] = `https://${this.config.assetsHost}/${this.config.clientId}/${value}`
-          }
-        } else {
-          accumulator[field.name] = value
-        }
+        accumulator[field.name] = resolveMediaRelativeToCloud(
+          value as string,
+          this.config
+        )
         break
       case 'rich-text':
         // @ts-ignore value is unknown
-        const tree = parseMDX(value, field)
+        const tree = parseMDX(value, field, this.config)
         accumulator[field.name] = tree
         break
       case 'object':