+++ experimental-examples/unit-test-example/CHANGELOG.md
@@ -4,6 +4,22 @@
 
 ### Patch Changes
 
+- Updated dependencies [999f0895a]
+- Updated dependencies [41be5e7fc]
+  - tinacms@0.68.8
+
+## null
+
+### Patch Changes
+
+- Updated dependencies [aaaa5bb09]
+- Updated dependencies [e06dbb3ca]
+  - tinacms@0.68.7
+
+## null
+
+### Patch Changes
+
 - Updated dependencies [2cc206b1a]
 - Updated dependencies [8998df207]
   - tinacms@0.68.6